+++ src/ArduinoJson/JsonDocument.hpp
@@ -18,9 +18,9 @@ class JsonDocument : public Visitable {
 
   JsonDocument() : nestingLimit(ARDUINOJSON_DEFAULT_NESTING_LIMIT) {}
 
-  template <typename T>
-  bool is() const {
-    return getVariant().template is<T>();
+  template <typename Visitor>
+  void accept(Visitor& visitor) const {
+    return getVariant().accept(visitor);
   }
 
   template <typename T>
@@ -33,28 +33,35 @@ class JsonDocument : public Visitable {
     return getVariant().template as<T>();
   }
 
-  template <typename T>
-  typename JsonVariantTo<T>::type to() {
-    _memoryPool.clear();
-    return getVariant().template to<T>();
-  }
-
   void clear() {
     _memoryPool.clear();
     _rootData.type = JSON_NULL;
+  }
+
+  template <typename T>
+  bool is() const {
+    return getVariant().template is<T>();
   }
 
   size_t memoryUsage() const {
     return _memoryPool.size();
   }
 
-  template <typename Visitor>
-  void accept(Visitor& visitor) const {
-    return getVariant().accept(visitor);
-  }
-
   TMemoryPool& memoryPool() {
     return _memoryPool;
+  }
+
+  template <typename T>
+  typename JsonVariantTo<T>::type to() {
+    _memoryPool.clear();
+    return getVariant().template to<T>();
+  }
+
+ protected:
+  template <typename T>
+  void copy(const JsonDocument<T>& src) {
+    nestingLimit = src.nestingLimit;
+    to<JsonVariant>().set(src.template as<JsonVariant>());
   }
 
  private:
@@ -76,13 +83,48 @@ class DynamicJsonDocument : public JsonDocument<DynamicMemoryPool> {
   DynamicJsonDocument(size_t capacity) {
     memoryPool().reserve(capacity);
   }
+
+  DynamicJsonDocument(const DynamicJsonDocument& src) {
+    memoryPool().reserve(src.memoryUsage());
+    copy(src);
+  }
+
+  template <typename T>
+  DynamicJsonDocument(const JsonDocument<T>& src) {
+    memoryPool().reserve(src.memoryUsage());
+    copy(src);
+  }
+
+  DynamicJsonDocument& operator=(const DynamicJsonDocument& src) {
+    copy(src);
+    return *this;
+  }
+
+  template <typename T>
+  DynamicJsonDocument& operator=(const JsonDocument<T>& src) {
+    copy(src);
+    return *this;
+  }
 };
 
 template <size_t CAPACITY>
 class StaticJsonDocument : public JsonDocument<StaticMemoryPool<CAPACITY> > {
  public:
+  StaticJsonDocument() {}
+
+  template <typename T>
+  StaticJsonDocument(const JsonDocument<T>& src) {
+    this->copy(src);
+  }
+
   StaticMemoryPoolBase& memoryPool() {
     return JsonDocument<StaticMemoryPool<CAPACITY> >::memoryPool();
+  }
+
+  template <typename T>
+  StaticJsonDocument operator=(const JsonDocument<T>& src) {
+    this->copy(src);
+    return *this;
   }
 };