+++ test/JsonDocument/StaticJsonDocument.cpp
@@ -6,9 +6,8 @@
 #include <catch.hpp>
 
 TEST_CASE("StaticJsonDocument") {
-  StaticJsonDocument<200> doc;
-
   SECTION("serializeJson()") {
+    StaticJsonDocument<200> doc;
     JsonObject obj = doc.to<JsonObject>();
     obj["hello"] = "world";
 
@@ -16,5 +15,88 @@ TEST_CASE("StaticJsonDocument") {
     serializeJson(doc, json);
 
     REQUIRE(json == "{\"hello\":\"world\"}");
+  }
+
+  SECTION("Copy assignment") {
+    StaticJsonDocument<200> doc1, doc2;
+    doc1.to<JsonVariant>().set(666);
+    deserializeJson(doc2, "{\"hello\":\"world\"}");
+    doc2.nestingLimit = 42;
+
+    doc1 = doc2;
+
+    std::string json;
+    serializeJson(doc1, json);
+    REQUIRE(json == "{\"hello\":\"world\"}");
+    REQUIRE(doc1.nestingLimit == 42);
+  }
+
+  SECTION("Copy constructor") {
+    StaticJsonDocument<200> doc1;
+    deserializeJson(doc1, "{\"hello\":\"world\"}");
+    doc1.nestingLimit = 42;
+
+    StaticJsonDocument<200> doc2 = doc1;
+
+    std::string json;
+    serializeJson(doc2, json);
+    REQUIRE(json == "{\"hello\":\"world\"}");
+    REQUIRE(doc2.nestingLimit == 42);
+  }
+
+  SECTION("Assign from StaticJsonDocument of different capacity") {
+    StaticJsonDocument<200> doc1;
+    StaticJsonDocument<300> doc2;
+    doc1.to<JsonVariant>().set(666);
+    deserializeJson(doc2, "{\"hello\":\"world\"}");
+    doc2.nestingLimit = 42;
+
+    doc1 = doc2;
+
+    std::string json;
+    serializeJson(doc1, json);
+    REQUIRE(json == "{\"hello\":\"world\"}");
+    REQUIRE(doc1.nestingLimit == 42);
+  }
+
+  SECTION("Assign from DynamicJsonDocument") {
+    StaticJsonDocument<200> doc1;
+    DynamicJsonDocument doc2;
+    doc1.to<JsonVariant>().set(666);
+    deserializeJson(doc2, "{\"hello\":\"world\"}");
+    doc2.nestingLimit = 42;
+
+    doc1 = doc2;
+
+    std::string json;
+    serializeJson(doc1, json);
+    REQUIRE(json == "{\"hello\":\"world\"}");
+    REQUIRE(doc1.nestingLimit == 42);
+  }
+
+  SECTION("Construct from StaticJsonDocument of different size") {
+    StaticJsonDocument<300> doc2;
+    deserializeJson(doc2, "{\"hello\":\"world\"}");
+    doc2.nestingLimit = 42;
+
+    StaticJsonDocument<200> doc1 = doc2;
+
+    std::string json;
+    serializeJson(doc1, json);
+    REQUIRE(json == "{\"hello\":\"world\"}");
+    REQUIRE(doc1.nestingLimit == 42);
+  }
+
+  SECTION("Construct from DynamicJsonDocument") {
+    DynamicJsonDocument doc2;
+    deserializeJson(doc2, "{\"hello\":\"world\"}");
+    doc2.nestingLimit = 42;
+
+    StaticJsonDocument<200> doc1 = doc2;
+
+    std::string json;
+    serializeJson(doc1, json);
+    REQUIRE(json == "{\"hello\":\"world\"}");
+    REQUIRE(doc1.nestingLimit == 42);
   }
 }